GET notes
Implemented api call to get real notes from api.

Implemented a useFetch hook to reuse.

Left a few TODOs as this took a while to get working and exceeded my time box

MUI seems to have an issue with the textField causing a resize observer error when multilines is used. https://github.com/mui/base-ui/issues/167

Downgraded mui to a version where this isn't an issue